refactor: Extract PathHelpers to a dedicated module and update imports
- Moved PathHelpers utility functions from the config module to a new path-helpers module for better organization and modularity.
- Updated import statements across various scripts to reflect the new PathHelpers location, ensuring consistent usage throughout the codebase.
- Removed obsolete default values and streamlined the config module, enhancing maintainability and clarity.
