+++ scripts/shared/validation-only-pipeline.ts
@@ -2,7 +2,8 @@ import { pipe } from 'remeda';
 import { existsSync } from 'fs';
 import type { Result, YamlFilesAndSchemasToWatch } from './validation-pipeline';
 import { validateYamlFilesAgainstSchemasPipeline } from './validation-pipeline';
-import { TAILOR_YAML_FILES_AND_SCHEMAS, COMPANY_FILES, PathHelpers } from './config';
+import { TAILOR_YAML_FILES_AND_SCHEMAS, COMPANY_FILES } from './config';
+import { PathHelpers } from './path-helpers';
 import { chain, tryCatch } from './functional-utils';
 
 // ============================================================================
